debian: update the debian configuration for packaging
Update to a 2.1.0 package and remove the specific 1.5.0 patches.

This sets up the debian framework to build a native package
directly out of the cloned repository. The debian directory
is moved out of the src dir to just avoid any patch complications
with and makes the quilt setup a bit easier for northbound
maintainers.

Signed-off-by: Charles Hardin <ckhardin@gmail.com>